modules/steam: Keep transient scope for gamescope child
This won't supervise the whole process tree, so what's started by
gamescope will live outside of the scope, but it fixes the issues with
being unable to ask logind to use the DRM display.

It should be sufficient for what is desired. Gamescope will quit once
the child process it manages goes away.