modules/steam: Add workarounds for OOBE
The steam updater will otherwise attempt to update even though there is
no network connection.

So we'll use approximately the vendor methods to decide what to do here.

Additionally, we'll always disable auto-updates when offline. This will
prevent stranding users with a valid steam install who happen to be
offline.