docs: Tautologically repeat the option name
***But Why?***

Simple! The header for the section is used as an anchor+link pair. It
makes it <del>awkward</del> really hard to copy the option name from it.

Instead of doing convoluted stuff to get it easier to copy from the
header, just add it again in there. It won't do any harm.