jovian-support-scripts: Init with format scripts
The support scripts are intended to replace vendor scripts for the
little details we can't do the same way.

We're adding the pair of scripts required to format external devices.
Note that the scripts end-up relying on the steamos-manager service,
which in turn already knows how to launch the hw-support scripts.
How convenient.