gamescope: Add Steam Big picture overlay synthesis patch
This patch set aims to make the Steam Overlay (“Steam Menu”) and the QAM
(“Quick Access Menu”) through additional bindings.

The bindings are the following:

 - F21 → Steam Overlay
 - F22 → QAM

The bindings were chosen so it would be unlikely to clash with any
keyboard input.

The bindings are expected to be used through external configuration,
changing a user-selected input into F21 or F22 to bring those menus.

For example, using udev and hwdb to remap a keyboard key.

Additionally, an in-gamescope method to bring-up the Steam Overlay was
added. This is needed since adding a system-wide remap for the keyboard
is undesirable, when this is a workaround for the lack of a feature
within Steam.