overlay: Guard against repeat overlaying
Multiple users have ended-up accidentally doing this, or the equivalent
of this in their configuration:

```nix
{
  imports = [ "${jovian}/modules" ];
  nixpkgs.overlays = [ jovian.overlays.default ];
}
```

When the overlay *overrides* an existing package, like for example the
`gamescope` package, this package gets the override applied twice.

In turn, if the override *appends* to the package attributes, the appended
values will be present twice.

Finally, when patches are attempted to be applied twice, all hell breaks
loose.

Fixes #486