overlay: Add internal flag to prevent guard clause application
This allows the documentation to re-import the overlay to peek at the
attribute names.

The documentation actually doesn't "re-overlay" the overlay in that
use-case, it only imports it to get the attribute names from the
overlay.

I guess another option to fix this would be to expose a
`__jovian_overlay_self` attribute on the overlay, which would in turn
contain only the "internal" overlay attribute set, rather than the whole
package set.

Though in reality, it might be best to consider this something that is
required in a new overlay "interface": a way that the overlay can
self-describe that it (1) exists, and (2) some of its properties.

In turn, generic tooling should be able to determine the overlays being
imported in a global `pkgs` set, and do *operations* on them.

With a proper interface, operations like exposing the overlay content
for documentation purpose should be trivial enough.

With enough care, it should also be possible to implement linting
externally, since one would have something tangible to work from.

Finally, with a proper interface, the overlaying system might be able to
warn and/or error on overlays being applied more than once. Though this,
too, must be done with care...